+++ src/main/java/com/fisa/study/management/domain/room/entity/Room.java
@@ -52,5 +52,4 @@ public class Room {
     @ManyToOne
     @JoinColumn(name = "member_id")
     private Member member;
-
 }
