+++ src/main/java/com/fisa/study/management/domain/snapshot/repository/SnapshotRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.UUID;
 
 public interface SnapshotRepository extends JpaRepository<Snapshot,Long> {
 
@@ -18,17 +19,17 @@ public interface SnapshotRepository extends JpaRepository<Snapshot,Long> {
 
 
 
-    @Query("select s from Snapshot s where s.room.id = :roomId " +
+    @Query("select s from Snapshot s where s.room.uuid = :uuid " +
             "and FUNCTION('MONTH', s.createdDate) =  :month " +
             "and FUNCTION('YEAR', s.createdDate) =  :year "+
             "and FUNCTION('DAY', s.createdDate) =  :day ")
-    List<Snapshot> findCreatedDateByRoomIdAndDay(@Param("roomId") Long roomId, @Param("year") int year ,
+    List<Snapshot> findCreatedDateByRoomUuidAndDay(@Param("uuid") UUID uuid, @Param("year") int year ,
                                                  @Param("month") int month, @Param("day") int day);
 
     @Query("SELECT s.createdDate FROM Snapshot s " +
-            "WHERE s.room.id = :roomId " +
+            "WHERE s.room.uuid = :uuid " +
             "AND FUNCTION('MONTH', s.createdDate) = :month " +
             "AND FUNCTION('YEAR', s.createdDate) = :year")
-    List<LocalDateTime> findDistinctCreatedDatesByRoomIdAndMonth
-            (@Param("roomId") Long roomId, @Param("year") int year,@Param("month") int month);
+    List<LocalDateTime> findDistinctCreatedDatesByRoomUuidAndMonth
+            (@Param("uuid") UUID uuid, @Param("year") int year, @Param("month") int month);
 }
