+++ src/main/java/com/fisa/study/management/domain/snapshot/service/SnapshotService.java
@@ -28,25 +28,18 @@ public class SnapshotService {
         return snapshotRepository.save(regSnapshotDTOToEntity(room, dto));
     }
 
-    public ResSnapshotDTO entityToSendSnapshotDTO(Snapshot snapshot) {
-        return ResSnapshotDTO.builder()
-                .title(snapshot.getTitle())
-                .content(snapshot.getContent())
-                .createdDate(snapshot.getCreatedDate())
-                .build();
-    }
 
     public Integer[] getSnapshotDateByDate(UUID uuid, int year,int month) {
-        Room room= roomRepository.findByUuid(uuid).orElseThrow(() -> new EntityNotFoundException("Room not found"));
+//        Room room= roomRepository.findByUuid(uuid).orElseThrow(() -> new EntityNotFoundException("Room not found"));
         return  snapshotRepository
-                .findDistinctCreatedDatesByRoomIdAndMonth(room.getId(),year,month)
+                .findDistinctCreatedDatesByRoomUuidAndMonth(uuid,year,month)
                 .stream().map(LocalDateTime::getDayOfMonth).distinct().toArray(Integer[]::new);
 
     }
 
     public List<ResSnapshotDTO> findSnapShotByRoomIdAndDay(UUID uuid,int year,int month,int day)  {
-        Room room= roomRepository.findByUuid(uuid).orElseThrow(() -> new EntityNotFoundException("Room not found"));
-        List<Snapshot> snapshots = snapshotRepository.findCreatedDateByRoomIdAndDay(room.getId(),year,month,day);
+//        Room room= roomRepository.findByUuid(uuid).orElseThrow(() -> new EntityNotFoundException("Room not found"));
+        List<Snapshot> snapshots = snapshotRepository.findCreatedDateByRoomUuidAndDay(uuid,year,month,day);
         return snapshots.stream().map(this::EntityToResSnapShotDTO).collect(Collectors.toList());
     }
 
@@ -67,4 +60,12 @@ public class SnapshotService {
         snapshot.setRoom(room);
         return snapshot;
     }
+    public ResSnapshotDTO entityToSendSnapshotDTO(Snapshot snapshot) {
+        return ResSnapshotDTO.builder()
+                .title(snapshot.getTitle())
+                .content(snapshot.getContent())
+                .createdDate(snapshot.getCreatedDate())
+                .build();
+    }
+
 }
